vMixTimecodeApp/requirements.txt
requests
Pillow
keyboard
pyperclip
aiohttp
PyInstaller
